fix(Service, Controller, Mapper)修复在添加、修改套餐和菜品的分类时能修改到不对应分类的问题

+++ sky-server/src/main/java/com/sky/controller/user/CategoryController.java
@@ -28,7 +28,7 @@ public class CategoryController {
      */
     @GetMapping("/list")
     @ApiOperation("分类查询")
-    public Result<List<Category>> list(Integer type){
+    public Result<List<Category>> list(Long type){
         List<Category> list = categoryService.getByListType(type);
         return Result.success(list);
     }
